# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documentation",
        "description": "This is the API documentation for the carpooling application.",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/trips": {
            "get": {
                "tags": [
                    "Trips"
                ],
                "summary": "Get all trips",
                "operationId": "692b0f6b729af25639ef76046f8fba1f",
                "parameters": [
                    {
                        "name": "start",
                        "in": "query",
                        "description": "Starting point of the trip",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "end",
                        "in": "query",
                        "description": "Ending point of the trip",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "date",
                        "in": "query",
                        "description": "Starting date of the trip",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Trip"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Trips"
                ],
                "summary": "Create a new trip",
                "operationId": "330c41008c88334c0e7653cbe4f0710a",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Trip"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Trip created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Trip"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/trips/{id}": {
            "get": {
                "tags": [
                    "Trips"
                ],
                "summary": "Get a trip by ID",
                "operationId": "6b8785740e1e7723d8899ea7bd468b48",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Trip"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Trip not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Trips"
                ],
                "summary": "Update a trip",
                "operationId": "92f9255028e296ba3829d9f96d7c37ed",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Trip"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Trip updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Trip"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Trip not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Trips"
                ],
                "summary": "Delete a trip",
                "operationId": "ec4541f129cf6d7a16feb2989476dd3a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Trip deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Trip not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Trip": {
                "title": "Trip",
                "required": [
                    "starting_point",
                    "ending_point",
                    "starting_at",
                    "available_places",
                    "price"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "readOnly": true,
                        "example": 1
                    },
                    "starting_point": {
                        "type": "string",
                        "example": "Paris"
                    },
                    "ending_point": {
                        "type": "string",
                        "example": "Lyon"
                    },
                    "starting_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-07-23T10:00:00Z"
                    },
                    "available_places": {
                        "type": "integer",
                        "example": 3
                    },
                    "price": {
                        "type": "integer",
                        "example": 20
                    },
                    "user_id": {
                        "type": "integer",
                        "example": 1
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
...
}